+++ src/app/page.tsx
@@ -41,26 +41,43 @@ export default function Home() {
     <>
       <Typography variant='body'>
         Hi there! I&apos;m an Associate Software Engineer currently working at{' '}
-        <ExternalLink href='https://www.ordnancesurvey.co.uk' message='Ordnance Survey' />, where I
-        help to build APIs and web solutions that underpin critical national services.
+        <ExternalLink
+          href='https://www.ordnancesurvey.co.uk'
+          message='Ordnance Survey'
+          noArrow={true}
+        />
+        , where I help to build APIs and web solutions that underpin critical national services.
       </Typography>
 
       <Typography variant='body'>
         <em className='pr-4'>Currently practising:</em>{' '}
-        <ExternalLink href='https://react.dev' message='React' />,{' '}
-        <ExternalLink href='https://tailwindcss.com/' message='Tailwind' />, and{' '}
-        <ExternalLink href='https://emotion.sh/docs/introduction' message='Emotion' /> on the
-        frontend, alongside geo utils such as{' '}
-        <ExternalLink href='https://maplibre.org/maplibre-gl-js/docs/' message='MapLibre' /> and{' '}
-        <ExternalLink href='https://turfjs.org/' message='Turf' />. For APIs and backend things
-        I&apos;m mostly using{' '}
-        <ExternalLink href='https://hub.docker.com/u/abiddiscombe' message='Docker' />,{' '}
-        <ExternalLink href='https://deno.com' message='Deno' />, and{' '}
-        <ExternalLink href='https://hono.dev' message='Hono' />, alongside{' '}
-        <ExternalLink href='https://mongodb.com' message='MongoDB' /> or{' '}
-        <ExternalLink href='https://postgresql.org' message='Postgres' />. I usually deploy to{' '}
-        <ExternalLink href='https://www.render.com' message='Render' />,{' '}
-        <ExternalLink href='https://azure.microsoft.com' message='Azure' />, or my home server.
+        <ExternalLink href='https://react.dev' message='React' noArrow={true} />,{' '}
+        <ExternalLink href='https://tailwindcss.com/' message='Tailwind' noArrow={true} />, and{' '}
+        <ExternalLink
+          href='https://emotion.sh/docs/introduction'
+          message='Emotion'
+          noArrow={true}
+        />{' '}
+        on the frontend, alongside geo utils such as{' '}
+        <ExternalLink
+          href='https://maplibre.org/maplibre-gl-js/docs/'
+          message='MapLibre'
+          noArrow={true}
+        />{' '}
+        and <ExternalLink href='https://turfjs.org/' message='Turf' noArrow={true} />. For APIs and
+        backend things I&apos;m mostly using{' '}
+        <ExternalLink
+          href='https://hub.docker.com/u/abiddiscombe'
+          message='Docker'
+          noArrow={true}
+        />
+        , <ExternalLink href='https://deno.com' message='Deno' noArrow={true} />, and{' '}
+        <ExternalLink href='https://hono.dev' message='Hono' noArrow={true} />, alongside{' '}
+        <ExternalLink href='https://mongodb.com' message='MongoDB' noArrow={true} /> or{' '}
+        <ExternalLink href='https://postgresql.org' message='Postgres' noArrow={true} />. I usually
+        deploy to <ExternalLink href='https://www.render.com' message='Render' noArrow={true} />,{' '}
+        <ExternalLink href='https://azure.microsoft.com' message='Azure' noArrow={true} />, or my
+        home server.
       </Typography>
 
       <Typography variant='body'>
@@ -69,6 +86,7 @@ export default function Home() {
         <ExternalLink
           href='https://www.durham.ac.uk/departments/academic/geography/'
           message='Durham University'
+          noArrow={true}
         />{' '}
         where I specialised in GIS and Glaciology.
       </Typography>