+++ src/components/SocialLink.tsx
@@ -6,7 +6,6 @@ export function SocialLink(p: SocialLinkProps) {
   const email = 'mailto://hello@abiddiscombe.dev';
   const github = 'https://github.com/abiddiscombe';
   const linkedin = 'https://linkedin.com/in/archiebiddiscombe';
-  const literature = 'https://github.com/abiddiscombe/literature';
 
   const genericClasses =
     'text-fontLighter-light dark:text-fontLighter-dark hover:text-font-light dark:hover:text-font-dark';
@@ -45,18 +44,6 @@ export function SocialLink(p: SocialLinkProps) {
           className={genericClasses}
         >
           <LinkedinIcon />
-        </a>
-      );
-
-    case 'literature':
-      return (
-        <a
-          href={literature}
-          target='_blank'
-          aria-label='Literature on GitHub (Opens in a New Tab)'
-          className={genericClasses}
-        >
-          <LiteratureIcon />
         </a>
       );
 
@@ -124,27 +111,6 @@ function LinkedinIcon() {
       <path d='M16 8a6 6 0 0 1 6 6v7h-4v-7a2 2 0 0 0-2-2 2 2 0 0 0-2 2v7h-4v-7a6 6 0 0 1 6-6z'></path>
       <rect x='2' y='9' width='4' height='12'></rect>
       <circle cx='4' cy='4' r='2'></circle>
-    </svg>
-  );
-}
-
-function LiteratureIcon() {
-  return (
-    <svg
-      xmlns='http://www.w3.org/2000/svg'
-      width='24'
-      height='24'
-      viewBox='0 0 24 24'
-      fill='none'
-      stroke='currentColor'
-      strokeWidth='2'
-      strokeLinecap='round'
-      strokeLinejoin='round'
-      className='feather feather-book h-6 md:h-5'
-    >
-      <title>Book Icon</title>
-      <path d='M4 19.5A2.5 2.5 0 0 1 6.5 17H20'></path>
-      <path d='M6.5 2H20v20H6.5A2.5 2.5 0 0 1 4 19.5v-15A2.5 2.5 0 0 1 6.5 2z'></path>
     </svg>
   );
 }