+++ src/components/blocks/Footer.tsx
@@ -1,0 +1,11 @@
+export function Footer() {
+  return (
+    <footer className='px-8'>
+      <div className='mx-auto max-w-3xl pt-12 pb-8'>
+        <small className='text-fontLighter-light dark:text-fontLighter-dark'>
+          &copy; Archie Biddiscombe
+        </small>
+      </div>
+    </footer>
+  );
+}